predict-stock-main/requirements.txt
flask
python-dotenv
pytz
tweepy
beautifulsoup4
lxml
transformers
tensorflow
pandas
numpy
gunicorn
torch
